Factor out Monad[Either[A, B]]
