+++ handlers/identity-backfill/src/main/scala/com/gu/identityBackfill/zuora/GetZuoraAccountsForEmail.scala
@@ -4,12 +4,10 @@ import com.gu.identityBackfill.Types._
 import com.gu.identityBackfill.salesforce.UpdateSalesforceIdentityId.IdentityId
 import com.gu.salesforce.TypesForSFEffectsData.SFContactId
 import com.gu.util.resthttp.Types._
-import com.gu.util.resthttp.Types.ClientSuccess
 import com.gu.util.zuora.SafeQueryBuilder.Implicits._
 import com.gu.util.zuora.ZuoraQuery.ZuoraQuerier
 import play.api.libs.json.Json
 import cats.implicits._
-import cats.data._
 
 object GetZuoraAccountsForEmail {
 
@@ -25,48 +23,15 @@ object GetZuoraAccountsForEmail {
   implicit val readsA = Json.reads[WireResponseAccount]
 
   def apply(zuoraQuerier: ZuoraQuerier)(emailAddress: EmailAddress): ClientFailableOp[List[ZuoraAccountIdentitySFContact]] = {
-
-    findZuoraContacts(zuoraQuerier, emailAddress)
-      .flatMap { list: List[WireResponseContact] =>
-        list.flatTraverse { (contactWithEmail: WireResponseContact) =>
-          val accountsM: ClientFailableOp[List[WireResponseAccount]] = findZuoraAccounts(zuoraQuerier, contactWithEmail)
-          val sfContactsM: ClientFailableOp[List[ZuoraAccountIdentitySFContact]] = accountsM.flatMap { accounts: List[WireResponseAccount] =>
-            accounts.traverse { accountsWithEmail: WireResponseAccount =>
-              clientFailableOpM.pure(ZuoraAccountIdentitySFContact(AccountId(accountsWithEmail.Id), accountsWithEmail.IdentityId__c.map(IdentityId.apply), SFContactId(accountsWithEmail.sfContactId__c), CrmId(accountsWithEmail.CrmId)))
-            }
+    findZuoraContacts(zuoraQuerier, emailAddress).flatMap {
+      _.flatTraverse { contactWithEmail =>
+        findZuoraAccounts(zuoraQuerier, contactWithEmail).flatMap {
+          _.traverse { accountsWithEmail =>
+            clientFailableOpM.pure(ZuoraAccountIdentitySFContact(AccountId(accountsWithEmail.Id), accountsWithEmail.IdentityId__c.map(IdentityId.apply), SFContactId(accountsWithEmail.sfContactId__c), CrmId(accountsWithEmail.CrmId)))
           }
-          accountsM.map { accounts: List[WireResponseAccount] =>
-            val tmp: List[ZuoraAccountIdentitySFContact] = accounts.map(accountsWithEmail => ZuoraAccountIdentitySFContact(AccountId(accountsWithEmail.Id), accountsWithEmail.IdentityId__c.map(IdentityId.apply), SFContactId(accountsWithEmail.sfContactId__c), CrmId(accountsWithEmail.CrmId)))
-            tmp.head
-          }
-          sfContactsM
         }
       }
-
-    //    findZuoraContacts(zuoraQuerier, emailAddress)
-    //      .flatMap { list: List[WireResponseContact] =>
-    //        list.traverse { (contactWithEmail: WireResponseContact) =>
-    //          (findZuoraAccounts(zuoraQuerier, contactWithEmail))
-    //            .map(_.map((accountsWithEmail: WireResponseAccount) => ZuoraAccountIdentitySFContact(
-    //              AccountId(accountsWithEmail.Id),
-    //              accountsWithEmail.IdentityId__c.map(IdentityId.apply),
-    //              SFContactId(accountsWithEmail.sfContactId__c),
-    //              CrmId(accountsWithEmail.CrmId)
-    //            )).flatten)
-    //        }
-    //      }
-
-    //    val accounts = for {
-    //      contactWithEmail <- findZuoraContacts(zuoraQuerier, emailAddress)
-    //      accountsWithEmail <- findZuoraAccounts(zuoraQuerier, contactWithEmail)
-    //    } yield ZuoraAccountIdentitySFContact(
-    //      AccountId(accountsWithEmail.Id),
-    //      accountsWithEmail.IdentityId__c.map(IdentityId.apply),
-    //      SFContactId(accountsWithEmail.sfContactId__c),
-    //      CrmId(accountsWithEmail.CrmId)
-    //    )
-    //
-    //    accounts.run
+    }
   }
 
   private def findZuoraAccounts(zuoraQuerier: ZuoraQuerier, contactWithEmail: WireResponseContact): ClientFailableOp[List[WireResponseAccount]] = {
